# test/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Kyber Control Plane API
  description: >
    Public HTTP API for the Kyber agent fleet control plane.
    All /api/v1/* endpoints require Bearer authentication.
    Webhook endpoints at /webhooks/* bypass Bearer auth.
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
paths:
...
  /api/v1/fleet/summary:
    get:
      operationId: getFleetSummary
      summary: Get fleet status summary
      tags: [fleet]
      responses:
        '200':
          description: Fleet summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/agents:
    get:
      operationId: listAgents
      summary: List all agents
      tags: [agents]
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createAgent
      summary: Create a new agent
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/agents/{name}:
    parameters:
      - $ref: '#/components/parameters/ResourceName'
    get:
      operationId: getAgent
      summary: Get an agent by name
      tags: [agents]
      responses:
        '200':
          description: Agent found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteAgent
      summary: Delete an agent
      description: >
        Permanently destroys the agent and all of its identity state (PVC,
        agent-scoped secrets, and external-store rows). Requires the
        confirmation interlock (?confirm={name}) and the lifecycle:admin scope
        (kyber#565). Irreversible until off-cluster backups (#171) land.
      tags: [agents]
      parameters:
        - $ref: '#/components/parameters/ConfirmName'
      responses:
        '204':
          description: Agent deleted
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    ResourceName:
      name: name
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9\-]{0,61}[a-z0-9]$'
        example: dave
    ConfirmName:
      name: confirm
      in: query
      required: true
      description: >
        Confirmation interlock (kyber#565). MUST equal the resource's {name}.
        A missing or mismatched value is rejected with 400 and nothing is
        deleted — guards against an accidental or automated destructive DELETE.
        BREAKING: this parameter is required; pre-#565 callers that omit it now
        receive 400.
      schema:
        type: string
        example: dave
  schemas:
...
    FleetSummary:
      type: object
      required: [machineCount, agentCount, agentsByPhase, machinesByPhase]
      properties:
        machineCount:
          type: integer
          description: Total number of machines
          example: 2
        agentCount:
          type: integer
          description: Total number of agents
          example: 5
        agentsByPhase:
          type: object
          description: Agent count grouped by phase
          additionalProperties:
            type: integer
          example:
            Running: 3
            Suspended: 2
        machinesByPhase:
          type: object
          description: Machine count grouped by phase
          additionalProperties:
            type: integer
          example:
            Running: 1
            Ready: 1
...
    AgentPhase:
      type: string
      enum:
        - Creating
        - Starting
        - Running
        - Stopping
        - Stopped
        - Restarting
        - Failed
        - Suspended
        - Deleted
        - ""
    AgentResources:
      type: object
      required: [cpu, memory, disk]
      properties:
        cpu:
          type: string
          example: "1"
        memory:
          type: string
          example: 2Gi
        disk:
          type: string
          example: 50Gi
    AgentStatus:
      type: object
      properties:
        phase:
          $ref: '#/components/schemas/AgentPhase'
        podName:
          type: string
        podIP:
          type: string
        restartCount:
          type: integer
          format: int32
        message:
          type: string
    Agent:
      type: object
      required: [id, phase, machine, runtime, authType, model, scaling, resources, createdAt]
      properties:
        id:
          type: string
          example: dave
        phase:
          $ref: '#/components/schemas/AgentPhase'
        machine:
          type: string
          example: worker-1
        runtime:
          type: string
          example: claude-code
        authType:
          type: string
          enum: [oauth, api-key]
        model:
          type: string
          example: claude-sonnet-4
        scaling:
          type: string
          enum: [warm, scale-to-zero]
          example: warm
        resources:
          $ref: '#/components/schemas/AgentResources'
        status:
          $ref: '#/components/schemas/AgentStatus'
        createdAt:
          type: string
          format: date-time
          example: "2026-04-10T00:00:00Z"
    AgentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
    CreateAgentRequest:
      type: object
      required: [name, machine, runtime, model]
      properties:
        name:
          type: string
          pattern: '^[a-z0-9][a-z0-9\-]{0,61}[a-z0-9]$'
          example: dave
        machine:
          type: string
          example: worker-1
        runtime:
          type: string
          example: claude-code
        model:
          type: string
          example: claude-sonnet-4
        scaling:
          type: string
          enum: [warm, scale-to-zero]
          default: warm
          example: warm
        resources:
          type: object
          properties:
            cpu:
              type: string
              example: "1"
            memory:
              type: string
              example: 2Gi
            disk:
              type: string
              example: 50Gi
        identity:
          type: object
          properties:
            soulDescription:
              type: string
        secrets:
          type: object
          properties:
            telegramEnabled:
              type: boolean
              default: false
            telegramBotToken:
              type: string
              writeOnly: true
            telegramAllowedUserIds:
              type: array
              items:
                type: string
            anthropicKey:
              type: string
            openaiApiKey:
              type: string
              writeOnly: true
            codexAuthJson:
              type: string
              description: Opaque auth.json from Codex ChatGPT login; write-only.
            authType:
              type: string
              enum: [oauth, api-key]
              default: oauth
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: name is required
        field:
          type: string
          example: name
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
  responses:
    ValidationError:
      description: Bad request — validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: >
        Authenticated but not authorized — the caller's API key lacks the scope
        this verb requires (kyber#474/#565). The body is intentionally non-leaky
        (it names neither the required scope nor the caller's granted scopes).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...